Time to start.  We're going to use our list of use cases as a guide for what to test next, and we'll do our best to keep this list sorted according to how simple each scenario actually is.  The simpler the scenario the higher in the list.
So, taking the first scenario, 'all cells are dead, should all
remain dead', we translate this into a Java method name.  Don't be
afraId of long test names.  At a minimum we want to encode two  pieces
of infomration in the name: the starting state of the system, and the
expected outcome.  If the scenario requires state transitions, include
that information in the name as well.

Do be careful about embedding implementation details in the name, like
the method being called. Automated IDE refactorings are not smart
enough to rename test names when methods names are changed.  More than
that, though, it implies that the test is coupled to the specific
implementation which is the exact opposite of what a good test is
supposed to be.