And passed!
I know, this feels like a cheating and a waste of time, but there's actually an important advancement hidden in this step. First, we are no longer hardcoding either the state of the individual grid cells or the size of the grid.  Second, the computed next generation is now derived from the input!

We've just completed one of the first and most basic steps along what Uncle Bob has termed the The Transformation Priority Premise (http://blog.cleancoder.com/uncle-bob/2013/05/27/TheTransformationPriorityPremise.html). We won't go into too much detail here, but this insight by Uncle Bob is a framework we will continue to utilize when maoking decision about what test to write next as we try to drive our Game of Life implementation toward a generic solution.

Using our use case notes, the next one up is 'All cells are alive', so let's start writer that test.