That's better! We're now forced to evaluate each cell.
You may argue that we could have made the assumption that only a single cell was alive and wrote code to identify the cell to kill. That's true, but I would argue that such an implementation would be less obvious and therefore not in the spirit of TDD.

Think back to the 3 steps:
Red - Write the simplest test that will fail
Green - Pass the test will the simplest solution
Refactor - Clean up or otherwise improve the solution

Our goal when passing the test is not to prove how clever we are but to write just enough code to get tphe new failing test passing, and to do so as quickly as we are able. Here, I feel the nested for loop is the most easiest solution to understand and therefore the simplest.

Now, with our code back in the "green" state lets see what our next test may be. We want to choose a test that will fail so we should consider a scenario that results in at least one cell being alive in the next generation. Review our notes, those scenarios are:

// live cell with two live neighbors, should live
// live cell with three live neighbors, should live
// dead cell with three live neighbors, should be alive

Thinking about these options, I think the "live cell with two live neighbors, should live" is the simplest, so lets start there.