Before we start refactoring lets take a moment to consider what makes the existing code difficult to understand. You probably immediately thought of the variables 'nx' and 'ny', which aren't great. There is also the problem of the 4 nested loops, at the bottom of which mutates a temporary variable. All of this is bad, but the reason for most of this is that our method is trying to do two things at: 1. count the number of alive neighbors and 2. change the next generation state based of that count.
So, rather than refactors at the smallest scale, we are actually going to first extract the alive neighbor counting logic into a dedicated method. We'll name this method something obvious like 'countAliveNeighbors' and provide all of the information necessary for it to accomplish the task.

Doing this yields immediate results. While we still have hte temporary
variable 'aliveNeighbors', we are no longer changing it within our outer
double for loop. This simple change is significantly easier to
understand, and significantly less prone to error.

Our new method 'countAliveNeighbors' is also very simple. The bad names
still exist, but there are zero side effects and it does exactly what
its name claims, a great trait for any bit of code!