In my humble opinion, this is much more straightforward. You may disagree, and there is definitely some problems around manually manipulating the indices, but think back to the Transformation Priority Premise we briefly discussed earlier. The idea of that premise was to provide a series of transformations to gradually change code from specific to generic. That is generally a good thing! However, one advantage of specific code is it's often easier to understand than the generic version, and code that is easy to understand should be one of our primary goals.
So, we're actually using the same principle set out by the Transformation Priority Premise, just in reverse! When you find yourself needing to modify a particuarly confusing bit of code, consider this idea, and explore ways to first simplify what is in front of you by moving up the transformations set out in the Transformation Priority Premise.