And the test passes... oops! No worries, however. While ideally all of our tests will fail at first, this is a normal situation to find yourself in so don't worry.
When this happens, take a step back to understand why? And to consider more deeply what scenario will can be converted into a failing test. If you can't think of one, congrats! You may be done! Otherwise, start writing that new test. For us, that test will be a scenario that causes a dead cell to be alive in the next generation. In our notes we jotted down "dead cell with three live neighbors, should be alive", so we'll start there.