org.opentest4j.AssertionFailedError: array contents differ at index [1][1], expected: <1> but was: <0> 	at io.github.tduncan.gameoflife.GameOfLifeTest.liveCellWithTwoLiveNeighborsShouldContinueLivingInNextGeneration(GameOfLifeTest.java:108)
Even after finding another scenario our test is still failing! What's happening?

Well, we actually have a bug in our solution! The Game of Life is intended for each generation to be derived from the previous generation, but for each generation to be immutable. The term immutable should be a big clue as to what's going on here. We are actually modifying the current generation, not producing a new generation! And, sometimes, those modifications cause future cells evaluated at a later point in time to be assigned the wrong state!

So, before to get our tests back to passing we need to address this problem now.