org.opentest4j.AssertionFailedError: array contents differ at index [1][1], expected: <1> but was: <0> 	at io.github.tduncan.gameoflife.GameOfLifeTest.gridContainingAliveCellsAndRemainsTheSameInTheNextGeneration(GameOfLifeTest.java:52)
We've fixed that problem, but we've uncovered another failing test. Before you become discouraged, take comfort in knowing that our test is correct, but our solution is no longer good enough! We actually need to implement another rule that we were able to cheat in the past, which is that a live cell should remain alive when it has three live neighbors. So let's do that!